+++ packages/app-project/src/screens/ClassifyPage/components/ClassifierWrapper/ClassifierWrapper.js
@@ -1,6 +1,6 @@
 import Classifier from '@zooniverse/classifier'
 import auth from 'panoptes-client/lib/auth'
-import { func, shape } from 'prop-types'
+import { func, string, shape } from 'prop-types'
 import React, { Component, useContext } from 'react'
 import asyncStates from '@zooniverse/async-states'
 import { logToSentry } from '@helpers/logger'
