+++ packages/app-project/src/screens/ClassifyPage/components/ClassifierWrapper/index.js
@@ -1,1 +1,1 @@
-export { default } from './ClassifierWrapperContainer'
+export { default } from './ClassifierWrapperConnector'
